fix dtype and atari extra

+++ tests/test_env_checker.py
@@ -8,7 +8,7 @@ from stable_baselines3.common.env_checker import check_env
 
 class ActionDictTestEnv(gym.Env):
     action_space = Dict({"position": Discrete(1), "velocity": Discrete(1)})
-    observation_space = Box(low=-1.0, high=2.0, shape=(3,), dtype=np.float32)
+    observation_space = Box(low=-1.0, high=2.0, shape=(3,), dtype=np.float64)
 
     def step(self, action):
         observation = np.array([1.0, 1.5, 0.5])
